refactor: rename index.tsx to ConfigsListFilters.tsx
